move the whole data arm into its own directory
