+++ src/data/index.test.ts
@@ -1,8 +1,8 @@
 import BigNumber from "bignumber.js";
 import { AssetType } from "stellar-sdk";
 
-import { EffectType } from "./constants/data";
-import { reframeEffect } from "./data";
+import { EffectType } from "../constants/data";
+import { reframeEffect } from "./";
 
 describe("reframeEffect", () => {
   const observerAccount = {
